Move auth service to the users service
