+++ src/main/kotlin/com/itmo/microservices/demo/users/api/controller/UserController.kt
@@ -1,6 +1,8 @@
 package com.itmo.microservices.demo.users.api.controller
 
 import com.itmo.microservices.demo.users.api.model.AppUserModel
+import com.itmo.microservices.demo.users.api.model.AuthenticationRequest
+import com.itmo.microservices.demo.users.api.model.AuthenticationResult
 import com.itmo.microservices.demo.users.api.model.RegistrationRequest
 import com.itmo.microservices.demo.users.api.service.UserService
 import io.swagger.v3.oas.annotations.Operation
@@ -8,6 +10,7 @@ import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.media.Content
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.security.SecurityRequirement
+import org.springframework.security.core.Authentication
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.web.bind.annotation.*
@@ -49,4 +52,28 @@ class UserController(private val userService: UserService) {
     )
     fun deleteCurrentUser(@Parameter(hidden = true) @AuthenticationPrincipal user: UserDetails) =
             userService.deleteUser(user)
+
+    @PostMapping("/auth")
+    @Operation(
+        summary = "Authenticate",
+        responses = [
+            ApiResponse(description = "OK", responseCode = "200"),
+            ApiResponse(description = "User not found", responseCode = "404", content = [Content()]),
+            ApiResponse(description = "Invalid password", responseCode = "403", content = [Content()])
+        ]
+    )
+    fun authenticate(@RequestBody request: AuthenticationRequest): AuthenticationResult =
+        userService.authenticate(request);
+
+    @PostMapping("/refresh")
+    @Operation(
+        summary = "Refresh authentication",
+        responses = [
+            ApiResponse(description = "OK", responseCode = "200"),
+            ApiResponse(description = "Authentication error", responseCode = "403", content = [Content()])
+        ],
+        security = [SecurityRequirement(name = "bearerAuth")]
+    )
+    fun refresh(authentication: Authentication): AuthenticationResult =
+        userService.refresh(authentication)
 }
