+++ src/main/kotlin/com/itmo/microservices/demo/users/api/model/AuthenticationRequest.kt
@@ -1,3 +1,3 @@
-package com.itmo.microservices.demo.auth.api.model
+package com.itmo.microservices.demo.users.api.model
 
 data class AuthenticationRequest(val username: String, val password: String)
