+++ src/main/kotlin/com/itmo/microservices/demo/users/api/model/AuthenticationResult.kt
@@ -1,3 +1,3 @@
-package com.itmo.microservices.demo.auth.api.model
+package com.itmo.microservices.demo.users.api.model
 
 data class AuthenticationResult(val accessToken: String, val refreshToken: String)
