+++ src/main/kotlin/com/itmo/microservices/demo/users/api/service/UserService.kt
@@ -1,7 +1,10 @@
 package com.itmo.microservices.demo.users.api.service
 
 import com.itmo.microservices.demo.users.api.model.AppUserModel
+import com.itmo.microservices.demo.users.api.model.AuthenticationRequest
+import com.itmo.microservices.demo.users.api.model.AuthenticationResult
 import com.itmo.microservices.demo.users.api.model.RegistrationRequest
+import org.springframework.security.core.Authentication
 import org.springframework.security.core.userdetails.UserDetails
 
 interface UserService {
@@ -9,4 +12,6 @@ interface UserService {
     fun registerUser(request: RegistrationRequest)
     fun getAccountData(requester: UserDetails): AppUserModel
     fun deleteUser(user: UserDetails)
+    fun authenticate(request: AuthenticationRequest): AuthenticationResult
+    fun refresh(authentication: Authentication): AuthenticationResult
 }
